# cargo-ship-backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/thresholds": {
      "post": {
        "operationId": "AlarmController_createThreshold",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateThresholdDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建阈值配置",
        "tags": [
          "告警管理"
        ]
      },
      "get": {
        "operationId": "AlarmController_findAllThresholds",
        "parameters": [
          {
            "name": "equipmentId",
            "required": false,
            "in": "query",
            "description": "设备ID",
            "schema": {
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "type": "string"
            }
          },
          {
            "name": "metricType",
            "required": false,
            "in": "query",
            "description": "监测指标类型",
            "schema": {
              "type": "string",
              "enum": [
                "vibration",
                "temperature",
                "pressure",
                "humidity",
                "speed",
                "current",
                "voltage",
                "power"
              ]
            }
          },
          {
            "name": "ruleStatus",
            "required": false,
            "in": "query",
            "description": "规则状态",
            "schema": {
              "type": "string",
              "enum": [
                "enabled",
                "disabled"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码（从1开始）",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询阈值列表",
        "tags": [
          "告警管理"
        ]
      }
    },
    "/api/thresholds/{id}": {
      "get": {
        "operationId": "AlarmController_findOneThreshold",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询阈值详情",
        "tags": [
          "告警管理"
        ]
      },
      "put": {
        "operationId": "AlarmController_updateThreshold",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateThresholdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新阈值配置",
        "tags": [
          "告警管理"
        ]
      },
      "delete": {
        "operationId": "AlarmController_deleteThreshold",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除阈值配置",
        "tags": [
          "告警管理"
        ]
      }
    },
    "/api/alarms": {
      "get": {
        "operationId": "AlarmController_findAllAlarms",
        "parameters": [
          {
            "name": "equipmentId",
            "required": false,
            "in": "query",
            "description": "设备ID",
            "schema": {
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "type": "string"
            }
          },
          {
            "name": "severity",
            "required": false,
            "in": "query",
            "description": "严重程度",
            "schema": {
              "type": "string",
              "enum": [
                "low",
                "medium",
                "high",
                "critical"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "处理状态",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "processing",
                "resolved",
                "ignored"
              ]
            }
          },
          {
            "name": "startTime",
            "required": false,
            "in": "query",
            "description": "开始时间（Unix时间戳毫秒）",
            "schema": {
              "example": 1700000000000,
              "type": "number"
            }
          },
          {
            "name": "endTime",
            "required": false,
            "in": "query",
            "description": "结束时间（Unix时间戳毫秒）",
            "schema": {
              "example": 1700086400000,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码（从1开始）",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询告警列表",
        "tags": [
          "告警管理"
        ]
      }
    },
    "/api/alarms/{id}": {
      "get": {
        "operationId": "AlarmController_findOneAlarm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询告警详情",
        "tags": [
          "告警管理"
        ]
      },
      "put": {
        "operationId": "AlarmController_updateAlarmStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlarmStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新告警状态",
        "tags": [
          "告警管理"
        ]
      }
    },
...
    "/api/statistics/alarms": {
      "get": {
        "description": "统计指定时间范围内的告警数据，按严重程度和处理状态分组",
        "operationId": "QueryController_getAlarmStatistics",
        "parameters": [
          {
            "name": "equipmentId",
            "required": false,
            "in": "query",
            "description": "设备ID（可选，不填则统计所有设备）",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "startTime",
            "required": true,
            "in": "query",
            "description": "开始时间（Unix时间戳，毫秒）",
            "schema": {
              "example": 1700000000000,
              "type": "number"
            }
          },
          {
            "name": "endTime",
            "required": true,
            "in": "query",
            "description": "结束时间（Unix时间戳，毫秒）",
            "schema": {
              "example": 1700086400000,
              "type": "number"
            }
          },
          {
            "name": "severity",
            "required": false,
            "in": "query",
            "description": "严重程度筛选（可选）",
            "schema": {
              "enum": [
                "low",
                "medium",
                "high",
                "critical"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统计成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmStatsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "告警统计",
        "tags": [
          "统计与导出"
        ]
      }
    },
...
    "/api/export/alarms": {
      "post": {
        "description": "根据查询条件导出告警记录为Excel或CSV文件",
        "operationId": "QueryController_exportAlarms",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportAlarmsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "导出成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "导出告警记录",
        "tags": [
          "统计与导出"
        ]
      }
    },
...
  },
  "info": {
    "title": "TDuCargoShipsManagement API",
    "description": "船舶货物管理系统 API 文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateThresholdDto": {
        "type": "object",
        "properties": {
          "equipmentId": {
            "type": "string",
            "description": "设备ID",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "metricType": {
            "type": "string",
            "description": "监测指标类型",
            "enum": [
              "vibration",
              "temperature",
              "pressure",
              "humidity",
              "speed",
              "current",
              "voltage",
              "power"
            ],
            "example": "temperature"
          },
          "upperLimit": {
            "type": "number",
            "description": "上限值（超过此值触发告警）",
            "example": 85.5
          },
          "lowerLimit": {
            "type": "number",
            "description": "下限值（低于此值触发告警）",
            "example": 10
          },
          "duration": {
            "type": "number",
            "description": "持续时间（毫秒），超过阈值并持续该时间后才触发告警",
            "example": 300000,
            "minimum": 0
          },
          "severity": {
            "type": "string",
            "description": "严重程度",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "example": "high"
          },
          "ruleStatus": {
            "type": "string",
            "description": "规则状态",
            "enum": [
              "enabled",
              "disabled"
            ],
            "default": "enabled"
          }
        },
        "required": [
          "equipmentId",
          "metricType",
          "duration",
          "severity"
        ]
      },
      "UpdateThresholdDto": {
        "type": "object",
        "properties": {
          "equipmentId": {
            "type": "string",
            "description": "设备ID",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "metricType": {
            "type": "string",
            "description": "监测指标类型",
            "enum": [
              "vibration",
              "temperature",
              "pressure",
              "humidity",
              "speed",
              "current",
              "voltage",
              "power"
            ],
            "example": "temperature"
          },
          "upperLimit": {
            "type": "number",
            "description": "上限值（超过此值触发告警）",
            "example": 85.5
          },
          "lowerLimit": {
            "type": "number",
            "description": "下限值（低于此值触发告警）",
            "example": 10
          },
          "duration": {
            "type": "number",
            "description": "持续时间（毫秒），超过阈值并持续该时间后才触发告警",
            "example": 300000,
            "minimum": 0
          },
          "severity": {
            "type": "string",
            "description": "严重程度",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "example": "high"
          },
          "ruleStatus": {
            "type": "string",
            "description": "规则状态",
            "enum": [
              "enabled",
              "disabled"
            ],
            "default": "enabled"
          }
        }
      },
      "UpdateAlarmStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "处理状态",
            "enum": [
              "pending",
              "processing",
              "resolved",
              "ignored"
            ],
            "example": "resolved"
          },
          "handleNote": {
            "type": "string",
            "description": "处理说明",
            "example": "已现场检查并修复",
            "maxLength": 500
          }
        },
        "required": [
          "status"
        ]
      },
...
      "AlarmStatsResponseDto": {
        "type": "object",
        "properties": {
          "totalCount": {
            "type": "number",
            "description": "告警总数",
            "example": 45
          },
          "groupBySeverity": {
            "type": "object",
            "description": "按严重程度分组统计",
            "example": {
              "low": 10,
              "medium": 20,
              "high": 12,
              "critical": 3
            }
          },
          "groupByStatus": {
            "type": "object",
            "description": "按处理状态分组统计",
            "example": {
              "pending": 5,
              "processing": 8,
              "resolved": 30,
              "ignored": 2
            }
          }
        },
        "required": [
          "totalCount",
          "groupBySeverity",
          "groupByStatus"
        ]
      },
...
      "ExportResponseDto": {
        "type": "object",
        "properties": {
          "downloadUrl": {
            "type": "string",
            "description": "下载链接",
            "example": "https://example.com/downloads/monitoring_data_20231114.xlsx"
          },
          "expiresAt": {
            "type": "number",
            "description": "链接过期时间（Unix时间戳，毫秒）",
            "example": 1700006399999
          }
        },
        "required": [
          "downloadUrl",
          "expiresAt"
        ]
      },
      "AlarmExportQueryDto": {
        "type": "object",
        "properties": {
          "equipmentId": {
            "type": "string",
            "description": "设备ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "severity": {
            "type": "string",
            "description": "严重程度",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "example": "high"
          },
          "status": {
            "type": "string",
            "description": "处理状态",
            "enum": [
              "pending",
              "processing",
              "resolved",
              "ignored"
            ],
            "example": "resolved"
          },
          "startTime": {
            "type": "number",
            "description": "开始时间（Unix时间戳，毫秒）",
            "example": 1700000000000
          },
          "endTime": {
            "type": "number",
            "description": "结束时间（Unix时间戳，毫秒）",
            "example": 1700086400000
          }
        }
      },
      "ExportAlarmsDto": {
        "type": "object",
        "properties": {
          "queryConditions": {
            "description": "查询条件",
            "allOf": [
              {
                "$ref": "#/components/schemas/AlarmExportQueryDto"
              }
            ]
          },
          "exportFormat": {
            "type": "string",
            "description": "导出格式",
            "enum": [
              "excel",
              "csv",
              "pdf"
            ],
            "example": "excel"
          }
        },
        "required": [
          "queryConditions",
          "exportFormat"
        ]
      },
...
    }
  }
}